+++ CHANGES.md
@@ -7,6 +7,8 @@ changes.
 
 ## Unreleased
 
+## 0.4.2
+
 - The OSM importer can now handle MultiPolygon boundaries for clipping
 - The final route feature has a new `full_path` property, with every snapped and freehand node